More last functions out of __init__.py
